chore: move util functions to separate file and document+test
